range-v3 port: Revert to hand-coded iterators for itext_iterator and otext_iterator.
The range-v3 implementation of basic_iterator doesn't yet support generation
of input and output iterators that satisfy post-increment requirements.  This
change re-implements itext_iterator and otext_iterator as hand-coded iterators
so that iterators that conform to C++11 requirements are provided.

This fixes https://github.com/tahonermann/text_view-range-v3/issues/2 and
https://github.com/tahonermann/text_view-range-v3/issues/3.